+++ src/js/theme.js
@@ -396,7 +396,14 @@ class Theme {
 
   initToc() {
     const $tocCore = document.getElementById('TableOfContents');
-    if ($tocCore === null) return;
+    if ($tocCore === null){
+      return;
+    }
+    const $toc = document.getElementById('toc-auto');
+    if ($tocCore.innerText === ''){
+      $toc.style.display = 'none';
+      return;
+    }
     if (document.getElementById('toc-static').getAttribute('kept') || this.util.isTocStatic()) {
       const $tocContentStatic = document.getElementById('toc-content-static');
       if ($tocCore.parentElement !== $tocContentStatic) {
@@ -410,7 +417,6 @@ class Theme {
         $tocCore.parentElement.removeChild($tocCore);
         $tocContentAuto.appendChild($tocCore);
       }
-      const $toc = document.getElementById('toc-auto');
       const $page = document.getElementsByClassName('page')[0];
       const rect = $page.getBoundingClientRect();
       $toc.style.left = `${rect.left + rect.width + 20}px`;