+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - :tada: Feat: add PWA support
 - :bug: Fix: add scroll-margin to headerlink
 - :bug: Fix: Check if content is empty beforing passing
+- :bug: Fix: toc-auto display error ([#21](https://github.com/Lruihao/FixIt/issues/21))
 - :fire: Feat: remove lib smooth-scroll ([#1](https://github.com/Lruihao/FixIt/issues/1))
 - :lipstick: Style: change scroll-behavior to smooth
 - :lipstick: Style: page image margin